src/requirements.txt
langchain
langchain-openai
langgraph
openai
python-dotenv
requests
streamlit
geopy
haversine
pandas
geopy
haversine
langchain_community
astrapy
chainlit